Just rediscovered this.
`try.c` contains a prototype new implementation of red-black trees. It's
not finished yet, but it is promising, and the logic of the red-black
balancing algorithm "all came rushing back to me" when I looked at the
code just now.

`try.c.bak` contains a precursor, but it has some different notes in it
I think so I'm preferring to keep it safe in version control.